Divide moonstone/Scrollable.useSpottable
Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
