+++ packages/moonstone/Scrollable/use.js
@@ -1,0 +1,31 @@
+import Spotlight, {} from '@enact/spotlight';
+import {useEffect, useRef} from 'react';
+
+const useTouch = (instance, props ,{uiRef}) => {
+	// const {
+
+	// } = instance.current;
+	// const {
+
+	// } = props;
+
+	const variables = useRef({
+
+	});
+
+	// useEffects
+
+	useEffect(() => {
+
+	}, []);
+
+	// functions
+
+	return {
+
+	};
+}
+
+export {
+	useTouch
+};
